Honor the 'xml:space' attribute on a <trans-unit> element
This works with a bit of magic: the default value for the 'xml:space' attribute is
defined in the schema, but XSL isn't aware of that. So we actually check with the
schema itself, and then either use the default value or the specific value set
on the <trans-unit> element.

This matches the descriptions in http://www.xmlplease.com/xml/xmlspace/ and
http://docs.oasis-open.org/xliff/v1.2/os/xliff-core.html#xml_space

Fixes #1