Allow forcing a specific language instead of relying on auto-detection
This plugin isn't bound to what the XLIFF actually says should be the source and/or
target language, but it can produce output files for any language code.

Being able to force a specific language makes it easier to centralize translations
for an application, while still having the separate elements use fake languages.